Update after Nov 2021 release
This update adds 7T MRI data for 184 participants, MEG data
for 51 participants, and further fixes for data files of 99
participants. Additions and updates for a total of 289 participants
are included, adding/changing a total of 2 million files and
tracking more than 25 TB in addition.

Due to changes in the HCP-openaccess S3 bucket, all download URLs for
all 17+ million files were updated, too.